feat: change BinFinder logic (#1038)
This PR changes the `BinFinder` algorithm that returns the connected bins used to build seeds. Instead of always returning the 8 surrounding bins, the user may define a vector containing a map of z bins in the top and bottom sets of space points for each middle space point bin. The algorithm will only search for SP in the bins defined in the map. In case no map is defined, the algorithm returns the 8 surrounding bins for top and bottom layers.
That will allow us to configure the top and bottom maps independently form each other and also to
return the bins based on geometrical assumptions as it is done by the ATLAS algorithm, for example:

```
//  1     2     3     4     5     6     7     8     9     10     11     z bin index
// -------------------------------------------------------------------> Z[mm]
// vector containing the map of z bins in the top layer
seedingCfg.binFinderConfigTop = {{1},{1, 2},{2, 3},{3, 4},{4, 5},{5, 6, 7},{7, 8},{8, 9},{9, 10},{10, 11},{11}};
// vector containing the map of z bins in the bottom layer
seedingCfg.binFinderConfigBottom = {{1, 2},{2, 3},{3, 4},{4, 5},{5, 6},{6},{6, 7},{7, 8},{8, 9},{9, 10},{10, 11}};
```

these vectors are passed to `BinFinder` by the `SeedingAlgorithm`.

@noemina