feat: Introduced new flags for how Boost, Eigen and VecMem would be set up. (2021.11.05.) (#1071)
Made it possible not to set them up at all. In case Acts is built as part of a "larger" project, like [traccc](https://github.com/acts-project/traccc). And that "larger" project wants to set these up according to its own liking.

At the same time fixed the capitalisation of the flags in the CI configuration. (`on` != `ON` in CMake lingo.)

I used the same naming here that I've set up in the "traccc projects" in the last weeks. See for instance:
  - https://github.com/acts-project/algebra-plugins/pull/27
  - https://github.com/acts-project/detray/pull/129

Making this kind of update will be vital for setting up the build of Acts in [traccc](https://github.com/acts-project/traccc) in the same way.