feat: Add Delegate type (#1059)
This type is intended as a sort-of replacement of `std::function` without the virtual function overhead of that type. This is ultimately meant to support the new extension mechanism for the KF and CKF, where these delegates will be used for the calibrator, smoother, and so on.

Delegate type that allows type erasure of a callable without allocation and with a single level of indirection. This type can support:
- a free function pointer
-  - a pointer to a member function alongside an instance pointer
Note: `Delegate` does not assume ownership of the instance. You need to ensure that the lifetime of the callable   instance is longer than that of the @c Delegate.
Currently `Delegate` only supports callables that are `const`.

The usage of this type looks like this for a free function:

```cpp
int sumImplementation(int a, int b) {
  return a + b;
}

Delegate<int(int, int)> sum;
sum.connect<&sumImplementation>();

int c = sum(2, 2); // = 4
```

and like this for a member pointer with instance:

```cpp
struct Subtractor {
  int x;

  int subtract(int v) const {
    return v - x;
  }
};

Delegate<int(int)> subtract;

Subtractor instance{5};

subtract.connect<&Subtractor::subtract>(&instance);

int y = subtract(10); // = 5
```