fix: Python examples/tests can be run from any directory (#1074)
Python scripts in  `Examples/Python/tests` and `Examples/Scripts/Python` should access config / data files from the acts source tree by using a path relative to the script's location (also in the source tree). They should not rely on the user's CWD being in the source tree.

This was already done in many cases. This PR fixes the remaining issues.

It is now possible to run `pytest` from the build directory as well as the source tree.