refactor!: Change the way (C)KF can be extended (#1062)
This PR changes the way the extension mechanism for the (C)KF work. It switches from templating them on all of these extension structs to using the `Delegate` type to define the interface.

At the same time this PR restructures the measurement selection and internal implementation of the `filter` method. Selection should now be more efficient  (in place as much as possible) and not make use of our variant `Measurement` type anymore, but operate directly on the track states.

This PR is dependent on

- #1059
- #1060 
- #1061
and includes their changes, so this PR's changeset will reduce once they are merged.

I'll have to do some manual rebasing to undo changes needed to make the other 3 PRs standalone, so I'll keep this WIP until that's completed.