fix: vertex constraints  (#1094)
This very small PR fixes the settings of position and covariance matrix of the vertex constraint in the VertexFitterAlgorithm.hpp. When I ran the Billoir fitter with the old settings, I got the following errors:

14:15:37    VertexFit      ERROR     Error in vertex fit with constraint.
14:15:37    VertexFit      ERROR     Numeric failure in calculation.

Looking at the FullBilloirVertexFitterTests.cpp in /Tests/UnitTests/Core/Vertexing, showed that the constraints had a dimension missing (and also wrong units), which resolved the issue.