feat: Add ostream-operator for IntersectionStatus (#1105)
This is the first part of my GSF pull-request series, where I often need to print the Intersection status for debugging. I split up this PR, since it is independent on the following ones.

Basically this pulls the nested `enum class Status` out of the `Intersection<DIM>` template struct to a `enum class IntersectionStatus` at namespace scope. The `IntersectionStatus` is broadcasted into the `Intersection<DIM>` with a `using Status = IntersectionStatus` to remain compatability.

This is done because otherwise the `Status`-enum would depend on the template-parameter `DIM` and thus the `operator<<` would also needed to be templated on `DIM` which makes things more complicated than they need to be.