fix: Added wait statements to kernels and changed the atomic operation to use VecMem (#1102)
While testing the Sycl seedfinder on Intel's Arctic Sound P card, we found  with @krasznaa that the code is missing some explicit `wait` statements in a few places that would result in waiting for the kernel to finish before executing later part of the code. 

This hasn't appeared before in our tests probably because, in none of the devices that we run the code on, the kernels were run asynchronously. 

This is now fixed with this PR. 
I also changed the outdated handling of the atomic operation to make use of `vecmem::atomic` which we discussed with @krasznaa already some time ago.