refactor!: stepper-state must not contain a stepSize object anymore (#1113)
This PR is in preparation for the MultiStepper PR #1110

Since a single stepSize isn't useful for a multi-component state, this removes the requirement to have such an object and add/modifies the accessors in the steppers. In particular it does:

Modify the stepper.setStepSize(...) function to allow it to replace calls to stepSize.update(...)`
This is done by adding an optional bool parameter to control wether or not release the stepSize. This defaults to true, so the default behaviour is unchanged, so I think this is not breaking.
This does also modify the previousStepSize in the stepper-states. But I think this is not problematic, since this variable seems to be used just in unit tests, which still run through as far as I can tell.
    state.stepping.stepSize.update(size, ConstrainedStep::aborter);
    // ->
    stepper.setStepSize(state.stepping, size, ConstrainedStep::aborter, false);
Add the stepper-method stepper.getStepSize(...) to the steppers. This does change the StepperConcept and thus is a breaking change I think, since custom steppers might not work anymore.
    state.stepping.stepSize.value(ConstrainedStep::aborter);
    // ->
    stepper.getStepSize(state.stepping, ConstrainedStep::aborter);
Use stepper.outputStepSize(...) instead of stepSize.toString() when necessary.
BREAKING CHANGE: The StepperConcept does now require a getStepSize(...) method.