feat: Allow setting reverse filter p threshold (#1103)
We previously didn't set up the reverse filtering at all in the examples track fitting algorithm, so the WARNING in #1101 and #1077 was never observed.

This PR exposes a momentum threshold variable through the config struct and into python. The pytests run it with a threshold of 0 (no reverse filtering at all) and 1 TeV (almost always reverse filtering), so we should at least see this warning here until #1101 is merged.