test: Refactor kalman-fitter test so it can be reused for other fitters (#1122)
This PR restructures the testing code for the `KalmanFitter` so that it can be reused for other fitters. This especially targets the GSF, but could also be useful for the GX2F #1099.

This is mainly done by putting the code inside a header file and templating the functions on `fitter_t`, `fitter_options_t` and `parameters_t`.

This should not change the behavior or the results of the tests.