feat: Make binning explicitly configurable in TGeo example (#1126)
This PR makes the surface binning in the TGeo Detector example explicitly configurable. In the json configuration a vector of bins for ```loc0``` and ```loc1``` is added for the negative, central and positive detector regions respectively. The binning must be set per layer, meaning that the number of layers must be known when configuring the detector building. Alternatively, a single binning entry of ```0``` can be used to trigger autobinning.
Also adds itk module splitter options to python bindings and enables tgeo json config dump from existing json config.