fix: Prevent FPE in AMVF if covariance is not invertible (#1142)
It's unclear to me how this covariance can end up not being invertible, but it has been observed when running the ACTS AMVF in an ATLAS online context.

We think it's safer to do so, and therefore discard the vertex one level up the call chain.

This is the backported version of #1142 to the `v15.x` release.