fix: Prevent FPE in AMVF if covariance is not invertible (#1142)
It's unclear to me how this covariance can end up not being invertible, but it has been observed when running the ACTS AMVF in an ATLAS online context. 

I briefly discussed with @baschlag whether returning `true`, i.e. treating the vertex as *merged* is correct, and we think it's safer to do so, and therefore discard the vertex one level up the call chain.

I'm going to cherry pick this PR into branch `release/v15.x` to create a bugfix tag for the `v15` release.