fix: Delegate checks exact compatibility of signatures (#1141)
As described in #1140, signatures are not checked for **exact** compatibility by `Delegate`.

This PR introduces an additional check where it will try to assign a given callable to an expected function pointer type generated from the signature given to `Delegate`. This assignment only succeeds if the signatures match exactly. This is enforced by a `static_assert`.

Fixes #1140

/cc @Corentin-Allaire