feat: MultiStepper for the GSF (#1110)
This PR adds mainly the class MultiEigenStepperLoop with some helper code.

It mainly adds new, independent code, the only case where it interferes with other propagation code is the stepper-concept.

The contained UnitTests cover most functionality of the MultiEigenStepperLoop and are built mainly as template-functions, so they can later be reused for the other MultiStepper-implementation (which will come later). The functionality is mostly tested against the EigenStepper.