test: Remove expected_failures from tests (#1116)
This can lead to unnoticed unit test failures, as was the case for the TrapezoidBounds test.

This PR removes all of these cases. In some spots they were present due to the tests asserting the outcome of an operation that was hard coded in surface stubs. I don't think this provides any meaningful insight anyway, so I assert the stub behavior instead (i.e. the test succeeds).