fix: Delete assignment operator for internal SPs (#1160)
The internal spacepoint class promises that it has an assignment operator, but this is never implemented. This means that if you are to write some code that would invoke the assignment operator, you'll run into esoteric linker errors. We cannot usefully implement an assignment operator for this class, because it has a reference member. As such, we simply delete this operator.