fix: Prevent unphysical number of phi bins (#1161)
The bug described in issue #1161 is caused by the creation of negative
or (near) infinite number of phi bins, which cause the internal
mechanisms of `std::vector` to crash. This effect is caused by the fact
that the number of phi neighbours is not properly initialized, causing
it to take on uninitialized values.

This commit does two things to prevent this issue. First of all, it adds
a sensible default value for the number of phi neighbours, ensuring that
it never takes on uninitialized values. Secondly, a sanity check is
added to the space point grid to ensure that the phi delta value has a
physically meaningful value; if this is not the case, an exception is
thrown.

Resolves #1161.