fix: Reset the stepper Jacobian in the KF smoothing (#1152)
This PR solves the low Pt issue that was occurring when using the KF and the smoothing. The Jacobian of the stepper is now reset before performing the last propagation from the last state to the perigee after the smoothing. 

This is the equivalent of replacing the `stepper.update` by a `stepper.resetState` but this implementation allows us to avoid computing the free parameters twice.