refactor: Add method to reset jacobian in stepper (#1171)
When trying to use the ATLAS stepper with the KF I noticed that it wouldn't work since we use to set the stepping state jacobian to a BoundMatrix::Identity() which isn't compatible with the ATLAS stepper.
This PR add a method to the stepper to set the jacobian to the identity, and thus solve this issue. 
Similarly I had to transform the update to a reset for the propagation to the Perigee as it relied on manually setting some matrix afterward (which was not compatible with the ATLAS stepper).