perf: Optimize Eigen usage in covariance engine
This commit optimizes some of the Eigen usage in the covariance engine,
specifically in the critical path for the propagation examples. The
first optimisation we make is to introduce a tiled matrix multiplication
method, which takes 2i×2j matrices, and performs four i×j
multiplications instead, which Eigen can optimize far more easily.
Secondly, we reduce the number of floating point operations performed by
working with smaller submatrices wherever possible.

On my machine, the following performance is achieved in the propagation
example before this patch: 53.555595 ms/event. After this patch, we take
43.750143 ms/event. This performance gain is independent from the
performance gain of #1181.