feat: Implement blocked matrix multiplication
Eigen uses a built-in cost model to determine how it should perform
certain matrix operations. In general, this works quite well, but there
are some edge cases where Eigen can struggle to output efficient code.
An especially apparent instance of this is when multiplying 8x8
matrices, which we do very commonly in operations such as stepping. In
these cases, we observe that Eigen emits code for generalized
matrix-matrix multiplication, rather than specialized code for small
matrices. This adds significant overhead to the matrix multiplication,
and it slows down some parts of our code a lot.

Our proposed solution is to implement a wrapper around Eigen's matrix
multiplication code that breaks a multiplication down into smaller
operations through a technique called blocking. A matrix of size 2ix2j
can be split into four matrices of sizes ixj, which can then be
multiplied and added to provide a correct multiplication of the parent
matrices. In this way, we can encourage Eigen to emit code for its
hand-optimized small-matrix operations, which should give speed-up in
some cases.

The operation is also supported for odd-sized matrices.