refactor!: Using non const InternalSpacePoint objects (#1196)
This PR tries to fix usage of mutable properties in https://github.com/acts-project/acts/pull/1168 and https://github.com/acts-project/acts/pull/1143.

Tagging: @paulgessinger @robertlangenberg @LuisFelipeCoelho 

BREAKING CHANGE: Remove assumption on constness of InternalSpacePoint. 
The need of this major change is to allow seed finding and filtering to evaluate and set properties of InternalSpacePoints as soon as they are candidates for triplet formation and seed quality evaluation.