fix: Bug when sorting the SP in SeedFinder (#1143)
In PR #1084 I mentioned that we were seeing a degradation in tracking efficiency when sorting the SP in cotangent of theta in the `SeedFinder`. However I realised that I was sorting only the `linCircleVec` vector in `SeedFinder` and the `compatBottomSP` and `compatTopSP` vectors should also be sorted. 

This PR solves the bug and now the efficiency and the output should be the same as without the sorting.

@paulgessinger @robertlangenberg @noemina