build: Remove bundled dependencies in favor of CMake FetchContent (#1177)
As discussed in the past and also already implemented for vecmem, this PR removes the bundled copies of 

- nlohmann_json
- pybind11
- dfelibs
- autodiff

and uses CMake's FetchContent to grab a copy. Also added is an option `ACTS_USE_SYSTEM_PYBIND11` to allow using an already installed copy. For nlohmann_json and autodiff this was already present, and I don't think it's necessary for dfelibs.