fix: Reduce memory usage of multiplication tests (#1200)
As mentioned in #1199, #1184 has ballooned the memory usage of the helper test build. This is because we are instantiating a lot of tests, generating a lot of function calls and types. The amount of tests is, to be honest, a little bit excessive. This commit reduces the number of tests, such that we only test MxN by NxP tests, where M,N,P is 3,3,3, 4,4,4, 8,8,8, and 8,7,4. This should still cover most of the edge cases and use cases.

Closes #1199.