feat: Reintroduce default bin finder constructor (#1197)
In Acts version 15.0.0, we see support being added for dynamic bins in the bin finder. However, this also means that we no longer have useful default constructors for this class. This has caused me a few issues in terms of backwards compatibility. I think it would be a good idea to reintroduce some sane default configuration for the bin finder, and I discussed with @LuisFelipeCoelho that having an empty list of _z_ neighbours with a single _φ_ neighbour would be a good default.