feat: Add compatibility method to seedfinder (#1198)
In the past, the seedfinder could be called with three spacepoint ranges, one for the lower, one for the middle, and one for the top spacepoints. In recent times, the seedfinder has been optimized impressively, but this has meant making the function call for the seeder quite a bit more complex; it now requires an output iterator, a state, and an extent for radii. For compatibility and ease-of-use reasons, I think it would be nice to have an "old-style" interface. This commit adds that.