feat: Gaussian Sum Filter (#1134)
Finally this is the PR which contains the GSF algorithm. It consists of the files necessary in Core and two UnitTests. There are no example executables or scripts contained, these will be part of a future PR.

The GSF passes all unit tests which are also used for the KalmanFitter, except for one, which does not makes sense for the GSF. The other test checks the component-merging and the KL-mixture-reduction.

Probably there should be added some more tests, but for the start I think this is ready for a first review.

This PR depends on #1110 (The MultiEigenStepperLoop)