feat: Custom z bin search order (#1169)
This PR introduces a custom z looping feature, 
which allows us to change the order in which `BinnedSPGroup` search for the Space Points.

In the Main branch the groups are sorted in **phi** and **z** in ascending order. But in Athena **z** follows the order within the vector `{1, 2, 3, 4, 11, 10, 9, 8, 6, 5, 7}`

This PR introduces this same custom looping approach with the vector `zBinsCustomLooping`. If it is not defined, the previous behaviour is preserved.

@noemina @paulgessinger @robertlangenberg