ci: Update macOS CI dependencies (#1228)
It seems like with the current macOS version on GH Actions and the latest tarball, DD4hep complains about a ROOT python version mismatch. My PR to DD4hep which adds flag to relax this behavior hasn't made it into a release yet, so right now I just regenerated the macOS deps to match the versions again.

Let's see if this works now.