refactor!: Use enum class for NavigationDirection (#1206)
This change is motivated by arithmetic between non-class enums and other types that are not the same enum is deprecated in C++20. I overload the operators we use these enums in to support the same use cases.

I think this is breaking since it's part of the public propagator interface.

BREAKING CHANGE: The `NavigationDirection` enum changes to `enum class`. The values are renamed `NavigationDiretcion::forward` -> `NavigationDirection::Forward` and `NavigationDirection::backward` -> `NavigationDirection::Backward`.