feat: change rMax invalid_argument to warning (#1227)
This PR allows for rMax in the grid to be different than the one used in `BinnedSPGroup`. Instead of throwing std::invalid_argument it just dumps a warning.

This is useful for ITk strip SPs. But in case we don't want to change the invalid_argument check, another option would be to add another parameter in the calculation of `numRBins = (config.rMax + config.beamPos.norm())` that allows for an increase or decrease of this value.

@noemina @robertlangenberg @paulgessinger