ci: Physmon follow up: enable failure again
Based on #1222

With #1222 in, we can have the python level tests and the perf
monitoring fail independently of one another. This PR enables that
again. Also bumps `histcmp` to v0.4.0 for a fix for missing legends in
some plots, and to include standalone PDFs in addition to the HTML
report.