ci: Clang tidy 2022 (#1208)
This adds a clang-tidy static analysis job to our CI. On GitHub Actions, I tested this to run for about 3h, which is clearly too long. 

I leverage the mechanism that I originally intended for the GPU CI to run this job on CERN GitLab. Here it only takes about 35min.

This is behind additional permission checks, where we maintain an allow-list of people who the CI acts for. Otherwise this check should remain neutral, allowing merging without overriding.

The CMake configuration should allow us to relatively flexibly define which checks to include in the report, and if we want we can even fail the build for others. I would propose we go about like this: 

1. This PR enables a number of checks, not all of which we intend to completely fix. 
2. We see how this behaves in practice. This shouldn't block anything from merging for now, other than due to compilation failures (i.e. clang-tidy warnings are reported but ignored)
3. If this seems to make sense, I'll add a restricted set of checks that will fail the build, and add a PR to fix these checks.

Thoughts?